# tests/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "AuditDecision": {
        "description": "What the trail records: an authorization outcome, or the outcome of a verification.",
        "enum": [
          "allow",
          "deny",
          "verified",
          "rejected"
        ],
        "title": "AuditDecision",
        "type": "string"
      },
...
      "BenchJobRecord": {
        "additionalProperties": false,
        "description": "Bench's evaluation job record, named for the document rather than for the module.\n\nBench and Studio each own a ``JobRecord``, and they are genuinely different things \u2014 one tracks\nan evaluation, the other a design study. Mounted in one process they collide in the OpenAPI\ndocument, and FastAPI disambiguates by prefixing the module path, which produces\n``astro_mine__bench__leaderboard___jobs__JobRecord`` and a generated client method returning a\ntype nobody can type.\n\nThe collision is an artifact of *composition* \u2014 it exists only because this distribution serves\nboth surfaces from one document \u2014 so it is resolved here rather than by renaming a platform type\nthat is unambiguous on its own. This subclass adds nothing: every field is inherited, so the\ndocument and the platform model cannot drift.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "job_id": {
            "minLength": 1,
            "title": "Job Id",
            "type": "string"
          },
          "result_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Id"
          },
          "status": {
            "$ref": "#/components/schemas/SubmissionStatus"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "BenchJobRecord",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "The Astro-Mine REST tier: the Hub registry API, the Studio API, Cloud's submission edge and the Bench leaderboard, behind one deployable.",
    "title": "Astro-Mine-API",
    "version": "<version>"
  },
  "openapi": "3.1.0",
  "paths": {
    "/bench/audit": {
      "get": {
        "description": "Query the authN/authZ + verification audit trail \u2014 ``audit:read``, admin-only (bench#29).\n\nThe queryable half of \"disputes are auditable\" (bench.md \u00a79): filter by who, what, and the\noutcome, newest first.",
        "operationId": "bench_audit_trail",
        "parameters": [
          {
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "in": "query",
            "name": "decision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AuditDecision"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision"
            }
          },
          {
            "in": "query",
            "name": "resource",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource"
            }
          },
          {
            "in": "query",
            "name": "submission_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Submission Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEvent"
                  },
                  "title": "Response Bench Audit Trail",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Trail",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/bench/jobs/{job_id}": {
      "get": {
        "operationId": "bench_get_job",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchJobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/bench/leaderboard/{scenario_id}": {
      "get": {
        "operationId": "bench_leaderboard",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LeaderboardEntry"
                  },
                  "title": "Response Bench Leaderboard",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Leaderboard",
        "tags": [
          "leaderboard"
        ]
      }
    },
    "/bench/leaderboard/{scenario_id}/scorecards": {
      "get": {
        "description": "The full per-metric leaderboard dataset View renders (bench.md \u00a76; RM-P1-BENCH-12).\n\nSame ranking as ``/bench/leaderboard/{scenario_id}`` but every row carries its complete\nscorecard with per-metric uncertainty, so View shows scorecards and bounds, not just the\nprimary.",
        "operationId": "bench_leaderboard_scorecards",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewLeaderboard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scorecards",
        "tags": [
          "leaderboard"
        ]
      }
    },
...
    "/bench/scenarios": {
      "get": {
        "description": "The scenario ids in the zoo \u2014 account-free, like every read path.",
        "operationId": "bench_list_scenarios",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Bench List Scenarios",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Scenarios",
        "tags": [
          "scenarios"
        ]
      },
      "post": {
        "description": "Publish a ScenarioSpec into the hosted catalog \u2014 ``scenario:author`` (bench#29).\n\nThe write surface of the Postgres/pgvector zoo catalog (bench#33): only a maintainer or an\nadmin may add to the commons' benchmark catalog, and the act is audit-logged. Returns 503 on\na deployment whose catalog is the read-only packaged filesystem zoo.",
        "operationId": "bench_author_scenario",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Spec",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Bench Author Scenario",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Author Scenario",
        "tags": [
          "scenarios"
        ]
      }
    },
...
  }
}